Move StartMessageBuilder to util.messages.start
